requirements.txt
streamlit
google-genai 
pillow